% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R, R/map.R
\name{basemap_options}
\alias{basemap_options}
\title{Available basemap tile providers}
\usage{
basemap_options

basemap_options
}
\description{
Character vector of basemap tile names suitable for
\code{tmap::tm_basemap()}.

Character vector of basemap tile names suitable for \code{tmap::tm_basemap()}.
}
\examples{
basemap_options
basemap_options
}
