% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{dft_age_breaks}
\alias{dft_age_breaks}
\alias{dft_age_labels}
\title{DfT standard age band breaks and labels}
\format{
\code{dft_age_breaks} is a numeric vector of 11 cut points;
\code{dft_age_labels} is a character vector of 10 band labels.
}
\usage{
dft_age_breaks

dft_age_labels
}
\description{
Age band definitions matching those used in DfT road casualty
publications (which differ from the bands included in the raw STATS19
casualty data).
}
\examples{
cut(c(8, 23, 67), breaks = dft_age_breaks, labels = dft_age_labels)
}
