% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_region_summarise_tag_costs}
\alias{map_region_summarise_tag_costs}
\title{Map TAG collision costs by region}
\usage{
map_region_summarise_tag_costs(
  region_sf,
  variable = "total_cost",
  plot_year,
  palette = "wes.zissou1",
  custom_breaks = c(0, 2, 6, 10, 15, 30, 60, 100, 200, 400),
  title = NULL,
  legend_title = "Value (£million)"
)
}
\arguments{
\item{region_sf}{An \code{sf} data frame of LA boundaries with cost data.}

\item{variable}{Character. Column name to plot.}

\item{palette}{Character. Colour palette name.}

\item{custom_breaks}{Numeric vector of class break values.}

\item{title}{Character. Map title.}

\item{legend_title}{Character. Legend title.}
}
\value{
A \code{tmap} object.
}
\description{
Creates a choropleth map of total TAG collision costs by Local Authority,
with an inset table showing the 10 LAs with the greatest reduction
potential.
}
\examples{
\dontrun{
tm <- map_region_summarise_tag_costs(la_costs_sf, variable = "total_cost",
                           palette = "wes.zissou1",
                           title = "TAG costs 2024")
}
}
