% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_crashes_to_osm}
\alias{match_crashes_to_osm}
\title{Match crashes to nearest OSM road segment}
\usage{
match_crashes_to_osm(osm_network_sf, crash_sf)
}
\arguments{
\item{osm_network_sf}{An \code{sf} object of OSM network data.}

\item{crash_sf}{An \code{sf} object of crash points.}
}
\value{
The crash \code{sf} object with an added \code{osm_id} column.
}
\description{
Assigns each crash record to the nearest driving network segment from OSM.
}
\examples{
\dontrun{
crashes_matched <- match_crashes_to_osm(osm_network_sf = osm_data,
                                        crash_sf = crashes)
}
}
