% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_lsoa_2021}
\alias{match_lsoa_2021}
\title{Match LSOA codes to 2021 equivalents}
\usage{
match_lsoa_2021(casualties = NULL, vehicles = NULL)
}
\arguments{
\item{casualties}{Optional casualty data frame with \code{lsoa_of_casualty}.}

\item{vehicles}{Optional vehicle data frame with \code{lsoa_of_driver}.}
}
\value{
The input data frame with added 2021 LSOA codes and names.
}
\description{
Matches casualty or vehicle LSOA codes to 2021 LSOA codes using official
lookup tables from the \code{geographr} package.
}
\examples{
\dontrun{
cas_lsoa21 <- match_lsoa_2021(casualties = my_casualties)
veh_lsoa21 <- match_lsoa_2021(vehicles = my_vehicles)
}
}
