% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_casualty_types}
\alias{summarise_casualty_types}
\title{Simplify casualty types}
\usage{
summarise_casualty_types(
  casualties,
  summary_type = c("short_name", "in_or_on")
)
}
\arguments{
\item{casualties}{Casualty data frame.}

\item{summary_type}{Character. One of \code{"short_name"} or \code{"in_or_on"}.}
}
\value{
Casualty data frame with simplified type column.
}
\description{
Maps detailed casualty types to shorter names or categories.
}
\examples{
\dontrun{
cas_short <- summarise_casualty_types(casualties, summary_type = "short_name")
}
}
